Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,7 +12,7 @@ int linear_search(int *array, size_t size, int value)
 	if (array != NULL)
 		for (size_t index = 0; index < size; index++)
 		{
-			printf("Value checked array[%ld] = [%d] \n", (size_t)index, array[index]);
+			printf("Value checked array[%ld] = [%d] \n", index, array[index]);
 			if (array[index] == value)
 			{
 				return (index);
